Propagate operation input/output in sorted way (#1427)
